# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is HireMeMaybe API service that provide data for HireMeMaybe web app",
        "title": "HireMeMaybe API service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/company/profile/banner": {
            "post": {
                "description": "Only file that smaller than 10 MB with .jpg, .jpeg, or .png extension is permitted",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Upload banner file for company",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Upload your banner file",
                        "name": "banner",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully upload banner",
                        "schema": {
                            "$ref": "#/definitions/model.Company"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not logged in as company, User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File size is larger than 10 MB",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "File extension is not allowed",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/company/profile/logo": {
            "post": {
                "description": "Only file that smaller than 10 MB with .jpg, .jpeg, or .png extension is permitted",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Upload logo file for company",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Upload your logo file",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully upload logo",
                        "schema": {
                            "$ref": "#/definitions/model.Company"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not logged in as company, User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File size is larger than 10 MB",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "File extension is not allowed",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/cpsk/profile/resume": {
            "post": {
                "description": "Only file that smaller than 10 MB with .pdf extension is permitted",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CPSK"
                ],
                "summary": "Upload resume file for CPSK",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Upload your resume file",
                        "name": "resume",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully upload resume",
                        "schema": {
                            "$ref": "#/definitions/model.CPSKUser"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not logged in as CPSK, User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File size is larger than 10 MB",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "File extension is not allowed",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/file/{id}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "File"
                ],
                "summary": "Retrieve dowloadable attachment",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of wanted file",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieve file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Given file id not found",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Fail to send file content",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "model.Application": {
            "type": "object",
            "required": [
                "post_id",
                "resume_id"
            ],
            "properties": {
                "answer": {
                    "$ref": "#/definitions/model.ApplicationAnswer"
                },
                "answer_id": {
                    "type": "integer"
                },
                "applied_at": {
                    "type": "string"
                },
                "cpsk_id": {
                    "description": "CPSKID references CPSKUser.UserID (uuid)",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post_id": {
                    "description": "PostID references JobPost.ID",
                    "type": "integer"
                },
                "resume_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ApplicationAnswer": {
            "type": "object",
            "properties": {
                "expected_salary": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "programming_languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "right_to_work": {
                    "type": "string"
                },
                "year_of_experience": {
                    "type": "integer"
                }
            }
        },
        "model.CPSKUser": {
            "type": "object",
            "properties": {
                "applications": {
                    "description": "List of job applications made by the CPSK user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                },
                "resume_id": {
                    "type": "integer"
                },
                "soft_skill": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "model.Company": {
            "type": "object",
            "properties": {
                "banner_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "job_post": {
                    "description": "JobPost holds the company's job posts",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.JobPost"
                    }
                },
                "logo_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "verified_status": {
                    "type": "string"
                }
            }
        },
...
        "model.JobPost": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "company_id": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "exp_lvl": {
                    "type": "string"
                },
                "expiring": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "post_time": {
                    "type": "string"
                },
                "req": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.PunishmentStruct": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "profile_picture": {
                    "type": "string"
                },
                "punishment": {
                    "$ref": "#/definitions/model.PunishmentStruct"
                },
                "tel": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utilities.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}