# internal/controller/file_handler.go
package controller

import (
	"HireMeMaybe-backend/internal/model"
	"HireMeMaybe-backend/internal/utilities"
	"errors"
	"fmt"
	"io"
	"log"
	"net/http"
	"path/filepath"
	"strings"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

// UploadResume function handles the process of uploading a resume file for a user and updating the
// user's information in the database.
// @Summary Upload resume file for CPSK
// @Description Only file that smaller than 10 MB with .pdf extension is permitted
// @Tags CPSK
// @Accept mpfd
// @Produce json
// @Param Authorization header string true "Insert your access token" default(Bearer <your access token>)
// @Param resume formData file true "Upload your resume file"
// @Success 200 {object} model.CPSKUser "Successfully upload resume"
// @Failure 400 {object} utilities.ErrorResponse "Invalid authorization header"
// @Failure 401 {object} utilities.ErrorResponse "Invalid token"
// @Failure 403 {object} utilities.ErrorResponse "Not logged in as CPSK, User is banned"
// @Failure 413 {object} utilities.ErrorResponse "File size is larger than 10 MB"
// @Failure 415 {object} utilities.ErrorResponse "File extension is not allowed"
// @Failure 500 {object} utilities.ErrorResponse "Database error"
// @Router /cpsk/profile/resume [post]
func (jc *JobController) UploadResume(c *gin.Context) {

	var cpskUser = model.CPSKUser{}

	user := utilities.ExtractUser(c)

	// Retrieve original profile from DB
	if err := jc.DB.Preload("User").Where("user_id = ?", user.ID.String()).First(&cpskUser).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to retrieve user information from database: %s", err.Error()),
		})
		return
	}

	rawFile, err := c.FormFile("resume")
	var maxBytesError *http.MaxBytesError
	if errors.As(err, &maxBytesError) {
		c.JSON(http.StatusRequestEntityTooLarge, utilities.ErrorResponse{
			Error: err.Error(),
		})
		return
	}

	if err != nil {
		c.JSON(http.StatusInternalServerError, utilities.ErrorResponse{
			Error: fmt.Sprintf("Failed to retrieve file: %s", err.Error()),
		})
		return
	}

	extension := strings.ToLower(filepath.Ext(rawFile.Filename))
	if extension != ".pdf" {
		c.JSON(http.StatusUnsupportedMediaType, utilities.ErrorResponse{
			Error: fmt.Sprintf("Unsupported file extension: %s", extension),
		})
		return
	}

	f, err := rawFile.Open()
	if err != nil {
		c.JSON(http.StatusInternalServerError, utilities.ErrorResponse{Error: "Cannot open file"})
		return
	}
	defer func() {
		if err := f.Close(); err != nil {
			log.Fatal("Failed to close file")
		}
	}()

	fileBytes, err := io.ReadAll(f)
	if err != nil {
		c.JSON(http.StatusInternalServerError, utilities.ErrorResponse{Error: "Cannot read file"})
		return
	}

	cpskUser.Resume.Content = fileBytes
	cpskUser.Resume.Extension = ".pdf"

	if err := jc.DB.Session(&gorm.Session{FullSaveAssociations: true}).Save(&cpskUser).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to update user information: %s", err.Error()),
		})
		return
	}

	c.JSON(http.StatusOK, cpskUser)
}

// companyUpload function handles process of reading files from company upload.
func (jc *JobController) companyUpload(c *gin.Context, fName string) (model.Company, []byte, string) {
	var company = model.Company{}

	u, _ := c.Get("user")
	if u == nil {
		c.JSON(http.StatusUnauthorized, gin.H{
			"error": "User information not provided",
		})
		return company, nil, ""
	}

	user, ok := u.(model.User)
	if !ok {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": "Failed to assert type",
		})
		return company, nil, ""
	}

	// Retrieve original profile from DB
	if err := jc.DB.Preload("User").Where("user_id = ?", user.ID.String()).First(&company).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to retrieve user information from database: %s", err.Error()),
		})
		return company, nil, ""
	}

	rawFile, err := c.FormFile(fName)
	var maxBytesError *http.MaxBytesError
	if errors.As(err, &maxBytesError) {
		c.JSON(http.StatusRequestEntityTooLarge, gin.H{
			"error": err.Error(),
		})
		return company, nil, ""
	}
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to retrieve file: %s", err.Error()),
		})
		return company, nil, ""
	}

	allowedExtensions := map[string]bool{
		".jpg":  true,
		".jpeg": true,
		".png":  true,
	}
	extension := strings.ToLower(filepath.Ext(rawFile.Filename))

	if !allowedExtensions[extension] {
		c.JSON(http.StatusUnsupportedMediaType, gin.H{
			"error": fmt.Sprintf("Unsupported file extension: %s", extension),
		})
		return company, nil, ""
	}

	f, err := rawFile.Open()
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Cannot open file"})
		return company, nil, ""
	}
	defer func() {
		if err := f.Close(); err != nil {
			log.Fatal("Failed to close file")
		}
	}()

	fileBytes, err := io.ReadAll(f)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Cannot read file"})
		return company, nil, ""
	}

	return company, fileBytes, extension
}

// UploadLogo function handles company's logo uploading and updating company profile in database.
// @Summary Upload logo file for company
// @Description Only file that smaller than 10 MB with .jpg, .jpeg, or .png extension is permitted
// @Tags Company
// @Accept mpfd
// @Produce json
// @Param Authorization header string true "Insert your access token" default(Bearer <your access token>)
// @Param logo formData file true "Upload your logo file"
// @Success 200 {object} model.Company "Successfully upload logo"
// @Failure 400 {object} utilities.ErrorResponse "Invalid authorization header"
// @Failure 401 {object} utilities.ErrorResponse "Invalid token"
// @Failure 403 {object} utilities.ErrorResponse "Not logged in as company, User is banned"
// @Failure 413 {object} utilities.ErrorResponse "File size is larger than 10 MB"
// @Failure 415 {object} utilities.ErrorResponse "File extension is not allowed"
// @Failure 500 {object} utilities.ErrorResponse "Database error"
// @Router /company/profile/logo [post]
func (jc *JobController) UploadLogo(c *gin.Context) {

	company, fileBytes, fileExtension := jc.companyUpload(c, "logo")

	if fileBytes == nil {
		return
	}

	company.Logo.Content = fileBytes
	company.Logo.Extension = fileExtension

	if err := jc.DB.Session(&gorm.Session{FullSaveAssociations: true}).Save(&company).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to update user information: %s", err.Error()),
		})
		return
	}

	c.JSON(http.StatusOK, company)
}

// UploadBanner function handles company's banner uploading and updating company profile in database.
// @Summary Upload banner file for company
// @Description Only file that smaller than 10 MB with .jpg, .jpeg, or .png extension is permitted
// @Tags Company
// @Accept mpfd
// @Produce json
// @Param Authorization header string true "Insert your access token" default(Bearer <your access token>)
// @Param banner formData file true "Upload your banner file"
// @Success 200 {object} model.Company "Successfully upload banner"
// @Failure 400 {object} utilities.ErrorResponse "Invalid authorization header"
// @Failure 401 {object} utilities.ErrorResponse "Invalid token"
// @Failure 403 {object} utilities.ErrorResponse "Not logged in as company, User is banned"
// @Failure 413 {object} utilities.ErrorResponse "File size is larger than 10 MB"
// @Failure 415 {object} utilities.ErrorResponse "File extension is not allowed"
// @Failure 500 {object} utilities.ErrorResponse "Database error"
// @Router /company/profile/banner [post]
func (jc *JobController) UploadBanner(c *gin.Context) {
	company, fileBytes, fileExtension := jc.companyUpload(c, "banner")

	if fileBytes == nil {
		return
	}

	company.Banner.Content = fileBytes
	company.Banner.Extension = fileExtension

	if err := jc.DB.Session(&gorm.Session{FullSaveAssociations: true}).Save(&company).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": fmt.Sprintf("Failed to update user information: %s", err.Error()),
		})
		return
	}

	c.JSON(http.StatusOK, company)
}

// GetFile function retrieves a file from the database and sends it as a downloadable attachment in
// the response.
// @Summary Retrieve dowloadable attachment
// @Tags File
// @Produce octet-stream
// @Param Authorization header string true "Insert your access token" default(Bearer <your access token>)
// @Param id path string true "ID of wanted file"
// @Success 200 {string} binary "Successfully retrieve file"
// @Failure 400 {object} utilities.ErrorResponse "Invalid authorization header"
// @Failure 401 {object} utilities.ErrorResponse "Invalid token"
// @Failure 403 {object} utilities.ErrorResponse "User is banned"
// @Failure 404 {object} utilities.ErrorResponse "Given file id not found"
// @Failure 500 {object} utilities.ErrorResponse "Fail to send file content"
// @Router /file/{id} [get]
func (jc *JobController) GetFile(c *gin.Context) {
	var file model.File
	id := c.Param("id")

	if err := jc.DB.First(&file, id).Error; err != nil {
		c.String(http.StatusNotFound, "File not found")
		return
	}

	// Set Content-Disposition with file name and extension
	c.Writer.Header().Set("Content-Disposition", "attachment; filename="+fmt.Sprint(file.ID)+file.Extension)
	c.Writer.Header().Set("Content-Type", "application/octet-stream")
	c.Writer.Header().Set("Content-Length", fmt.Sprint(len(file.Content)))

	// Write file data to response
	_, err := c.Writer.Write(file.Content)
	if err != nil {
		if !c.Writer.Written() {
			c.JSON(http.StatusInternalServerError, gin.H{
				"error": "Failed to send file content",
			})
		} else {
			c.Abort()
		}
		return
	}
}